fix: 관객 앱 빌드 오류 수정
