+++ apps/spectator/app/page.tsx
@@ -37,7 +37,7 @@ const GAMES: Games[] = [
 ];
 
 type PageProps = {
-  searchParams: { [key: string]: string | string[] | undefined };
+  searchParams: Promise<{ [key: string]: string | string[] | undefined }>;
 };
 
 export default async function Page({ searchParams }: PageProps) {
