+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "hcc",
   "private": true,
+  "type": "module",
   "scripts": {
     "build": "turbo build --remote-only",
     "dev": "turbo dev --parallel",
@@ -15,6 +16,7 @@
   },
   "devDependencies": {
     "@commitlint/cli": "^19.6.1",
+    "@commitlint/types": "^19.5.0",
     "@commitlint/config-conventional": "^19.6.0",
     "husky": "^9.1.7",
     "lint-staged": "^15.4.3",
