+++ packages/styles/package.json
@@ -4,7 +4,6 @@
   "private": true,
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "sideEffects": false,
   "type": "module",
   "scripts": {
     "lint": "eslint . --max-warnings 0",
@@ -16,17 +15,16 @@
       "eslint --fix --max-warnings 0"
     ]
   },
+  "dependencies": {
+    "@vanilla-extract/css": "^1.17.1"
+  },
   "devDependencies": {
     "@hcc/eslint-config": "workspace:*",
     "@hcc/typescript-config": "workspace:*",
-    "@turbo/gen": "^1.13.4",
     "@types/node": "^22.13.0",
     "@types/react": "^19.0.8",
     "@types/react-dom": "^19.0.3",
     "react": "^19.0.0",
     "typescript": "^5.7.3"
-  },
-  "dependencies": {
-    "@vanilla-extract/css": "^1.17.1"
   }
 }
