+++ packages/ui/src/modal/Trigger.tsx
@@ -1,24 +1,17 @@
-import { ElementType, forwardRef } from 'react';
+import { ComponentProps } from 'react';
 
 import { useModal } from './hooks';
-import {
-  PolymorphicComponentProps,
-  PolymorphicComponentPropsWithRef,
-  PolymorphicRef,
-} from './type';
 
-type Props<C extends ElementType> = PolymorphicComponentProps<C>;
-type TriggerType = <C extends ElementType = 'button'>(
-  props: PolymorphicComponentPropsWithRef<C, Props<C>>,
-) => React.ReactNode;
+type ModalTriggerProps = {
+  onClick?: () => void;
+} & ComponentProps<'button'>;
 
-const ModalTrigger: TriggerType = forwardRef(function ModalTrigger<
-  C extends ElementType = 'button',
->(
-  { as, className, children, onClick, ...props }: Props<C>,
-  ref: PolymorphicRef<C>,
-) {
-  const Component = as || 'button';
+const ModalTrigger = ({
+  onClick,
+  className,
+  children,
+  ...props
+}: ModalTriggerProps) => {
   const { onOpenToggle } = useModal();
 
   const handleClick = () => {
@@ -27,10 +20,10 @@ const ModalTrigger: TriggerType = forwardRef(function ModalTrigger<
   };
 
   return (
-    <Component ref={ref} className={className} {...props} onClick={handleClick}>
+    <button className={className} {...props} onClick={handleClick}>
       {children}
-    </Component>
+    </button>
   );
-});
+};
 
 export default ModalTrigger;
